Fix missing 64-bit offset types in diskio-unix
Most of the types used in the diskio routines are explicitly 64-bit, but
there were two usages of off_t, which is a 32-bit value when compiled
in a 32-bit userspace.  This causes gpt-fdisk to be unable to correctly
write partitions on a drive larger than 4GiB when compiled in a 32-bit
userspace.  This change updates the two usages of off_t to off64_t,
which is explicitly 64-bit and allows the program to work as intended.

Signed-off-by: xlfeng <xulifeng7@huawei.com>